requirements.txt
Django>=2.2.2,<2.3.0
pyTelegramBotAPI>=3.6.6,<3.7.0
lxml>=4.4.1,<4.5.0
IMDbPY>=6.8,<7.0